Move the rest of createCard functions into index.ts card component
